Split events module into multiple files
This allows to keep things more organized as subscribers grow and also
break circular imports.
