+++ lms/app.py
@@ -101,7 +101,6 @@ def create_app(global_config, **settings):  # pylint: disable=unused-argument
     config.include("lms.services.jwt")
     config.include("lms.validation")
     config.include("lms.tweens")
-    config.include("lms.events")
     config.add_static_view(name="export", path="lms:static/export")
     config.add_static_view(name="static", path="lms:static")
